Expose require.dumpCache() for resetting the module cache.
Resetting the module cache is an important point of hygiene when running
multiple tests, and sometimes within a single test, because it prevents
earlier test activity from interfering with current test behavior.

Without modifying prelude.js, it seems to be impossible to modify the
cache in any way. Unfortunately, shipping a modified version of prelude.js
requires forking both browserify and browser-pack.

The other option would be to use something other than browserify for
tests, but then the benefits of end-to-end testing the actual browserified
bundle are lost.

Am I missing an easier way of clearing the cache?